use postinstall script for expo version
