+++ src/index.ts
@@ -1,6 +1,5 @@
 import type { ExpoConfig } from '@expo/config-types';
 
-import { withAnalytics } from './analytics/injectAnalytics';
 import { withCIOAndroid } from './android/withCIOAndroid';
 import { withCIOIos } from './ios/withCIOIos';
 import type { CustomerIOPluginOptions } from './types/cio-types';
@@ -17,8 +16,6 @@ function withCustomerIOPlugin(
   if (props.android) {
     config = withCIOAndroid(config, props.android);
   }
-
-  config = withAnalytics(config, props);
 
   return config;
 }
