--- src/version.ts
@@ -1,1 +1,0 @@
-export const LIB_VERSION = "1.0.0-alpha.3";
